Silence warnings about types, mostly in the test codes.

+++ source/test/LoadBalanceCorrectness/main-lbcorrectness.C
@@ -1617,7 +1617,7 @@ void enforceNesting(
 
    const hier::BoxLevel &L0 = L1_to_L0.getHead();
 
-   const long int cell_count = L1.getGlobalNumberOfCells();
+   const size_t cell_count = L1.getGlobalNumberOfCells();
 
    /*
     * Make L1 nest inside L0 by nesting_width.
